requirements.txt
Django==2.1
django-crispy-forms==1.7.2
djangorestframework==3.8.2
pytz==2018.5
